feat: implementation permissions by role

+++ app/crud/todo_list.py
@@ -12,10 +12,13 @@ def read_todo_list(
     email: Optional[str] = None,
     skip: int = 0,
     limit: int = 10,
-    db: Session = Depends(get_db_session)
+    db: Session = Depends(get_db_session),
+    current_user: dict = Depends()
 ):
     query = select(TodoList.id, TodoList.title, TodoList.description, User.username, TodoList.created_at).join(User, TodoList.owner_id == User.id)
 
+    if current_user["role"] in ["user"]:
+        query = query.where(User.username == current_user["sub"])
     if id:
         query = query.where(TodoList.id == id)
     if owner_id:
@@ -48,13 +51,20 @@ def read_todo_list(
 
     return new_todo_lists
 
-def create_todo_list(todo_list_create: TodoListCreate, db: Session = Depends(get_db_session)):
+def create_todo_list(
+    todo_list_create: TodoListCreate, 
+    db: Session = Depends(get_db_session), 
+    current_user: dict = Depends()
+):
     # Validar que el owner_username existe en la base de datos de usuarios
     owner_query = select(User).where(User.username == todo_list_create.owner_username)
     owner = db.exec(owner_query).first()
 
     if owner is None:
         raise HTTPException(status_code=404, detail="Owner not found")
+    
+    if current_user["role"] in ["user"] and current_user["sub"] != owner.username:
+        raise HTTPException(status_code=403, detail="Insufficient permissions to create todo list to other users")
     
     try:
         new_todo_list = TodoList(
@@ -82,7 +92,8 @@ def create_todo_list(todo_list_create: TodoListCreate, db: Session = Depends(get
 def update_todo_list(
     id: int,
     todo_list_update: TodoListUpdate,
-    db: Session = Depends(get_db_session)
+    db: Session = Depends(get_db_session),
+    current_user: dict = Depends()
 ):
     query = select(TodoList).where(TodoList.id == id)
     todo_list = db.exec(query).first()
@@ -95,6 +106,9 @@ def update_todo_list(
 
     owner_query = select(User).where(User.id == todo_list.owner_id)
     owner = db.exec(owner_query).first()
+
+    if current_user["role"] in ["user"] and current_user["sub"] != owner.username:
+        raise HTTPException(status_code=403, detail="Insufficient permissions to update todo list to other users")
 
     try:
         db.add(todo_list)
@@ -114,12 +128,22 @@ def update_todo_list(
 
     return returned_new_list
 
-def delete_todo_list(id: int, db: Session = Depends(get_db_session)):
+def delete_todo_list(
+    id: int, 
+    db: Session = Depends(get_db_session), 
+    current_user: dict = Depends()
+):
     query = select(TodoList).where(TodoList.id == id)
     todo_list = db.exec(query).first()
 
+    owner_query = select(User).where(User.id == todo_list.owner_id)
+    owner = db.exec(owner_query).first()
+
     if not todo_list:
         raise HTTPException(status_code=404, detail="Todo list not found")
+    
+    if current_user["role"] in ["user"] and current_user["sub"] != owner.username:
+        raise HTTPException(status_code=403, detail="Insufficient permissions to update todo list to other users")
     
     try:
         db.delete(todo_list)
